fix shard tracking issue for localnet (#6406)
Right now, localnet use a validator seat config that will split validators to all shards, as a result, every validator will track only a subset of shards in the localnet, even when ChunkOnly producer is not enabled. This is not desired since we want to run localnet similar to the setup of testnet and mainnet right now, where all validators track all shards. This PR changes that by changing how the localnet genesis config is generated.