nearcore: add IS_ARCHIVE key to the database (#6730)
Since archival nodes no longer generate values in ColTrieChanges
column, it is now ill-advised to switch a node from being an archival
to non-archival.  The node will work but will leave some garbage in
its database.

To prevent users from doing the switch, introduce IS_ARCHIVE key to
the database which indicates whether the database has been used with
an archival node.

If non-archival node opens such database it fails with an error.  An
archival node, on the other hand, will make sure that the key is set.

Fixes: https://github.com/near/nearcore/issues/6522