refactor: don't type-erase IO errors (#6806)
This changes from `IOErr(String)` to `IOErr(io::Error)`, so we don't
lose specific error. The consequence here is that Error is no longer Eq,
so the tests are adjusted accordingly.

In general, errors should be neither Eq nor Clone, they are not values.

Often times, when an error feels like a value, it means that the error
type wants to be split in two, from `Result<T, Error>` into
`Result<Result<T, InvalidTx>, io::Error>` where the outer Error captures
real errors/exceptions (such as the database having a bad day), and
internal error is just a normal outcome of a request processing.