refactor: make blocks and headers cheaply clonable (#6866)
Prep work for #6811

This is a somewhat large change of the internal model, as now this types
are mostly immutable. This seems reasonable -- I can imagine us doing
things like fetching a block on the database thread, putting it into the
cache and handling a reference to it to HTTP server.

It seems that blocks are immutable in practice -- the only exception are
some testing utilities, but `Arc::make_mut` deals with those cases.