runtime: move closures out of check_balance to refactor it slightly (#7263)
check_balance function is a pretty long one.  Attempt to shorten it
a little by moving closures out of it and converting them to
functions.  This does require a few more arguments to be passed to
those functions but I’d argue that’s beneficial since now it’s clear
what data each of the functions actually use without having to
analyse it carefully.