chain: maybe_block_id_to_block_hash → maybe_block_id_to_block_header (#7373)
In all instances where ViewClientActor::maybe_block_id_to_block_hash
is called, the caller takes the hash and fetches block’s header.
However, in two out of three cases the method already fetches the
header of the block (to determine the hash) making it trivially
available to the caller.

Replace the method with one which returns the header directly.  This
means that in one case where the fetching did not occur inside of the
method, it is now happening there.