sandbox: refactor SandboxStatePatch making it unit on non-sandbox builds (#7372)
Using conditional compilation, create a unit type definition of
a SandboxStatePatch struct.  With this approach it’s encoded in the
type system that only sandbox build can have non-empty state patch.

It also improves optimisation since the compiler can resolve at
compile-time that the patch is empty thus eliminating dead code as
well as reducing the size of the type from three pointers to nothing.